# Conflicts:
#	app/src/main/java/ca/dal/csci3130/quickcash/jobmanagement/Job.java

+++ app/src/main/java/ca/dal/csci3130/quickcash/jobmanagement/FeedbackActivity.java
@@ -33,7 +33,7 @@ public class FeedbackActivity extends AppCompatActivity {
     TextView userID;
     RatingBar ratingBar;
 
-    double myRating;
+    float myRating;
 
 
     @Override
